socket: set ipv6 preference to false
Torch relies on explicitly setting the address family to AF_INET6 by
default when using `getaddrinfo`. This results in ENONAME on systems
which do not have IPv6 available, as is the case within EKS in our setup.

The correct behavior should be to set the family to AF_UNSPEC, and if
necessary to rely on `gai.conf` to set precedence to control which
address is returned first. This will be addressed separately by raising
a PR upstream.

In the interval, our fork should default to IPv4 to allow host lookups
to work as intended.